pyproject.toml
[tool.poetry]
name = "rf-appium"
version = "0.1.0"
description = "Automação Mobile Com Appium e Robot Framework"
authors = ["Vinicius Trindade <vinicius.trindade@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
robotframework = "^7.0"
robotframework-appiumlibrary = "2.0.0"
appium-python-client = "3.2.1"
docutils = "0.20.1"
PyYAML = "6.0.1"
requests = "^2.32.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
